ref(omit+changes): heavy refactoring
